+++ src/netbsd/device.c
@@ -104,7 +104,7 @@ int demi_device_get_devunit(struct demi_device *dd, uint32_t *devunit)
     }
 
     if (dd->devunit != -1) {
-        *devunit = dd->devunit;
+        *devunit = (uint32_t)dd->devunit;
         return 0;
     }
 
@@ -115,7 +115,7 @@ int demi_device_get_devunit(struct demi_device *dd, uint32_t *devunit)
     for (i = 0; devname[i] != '\0'; i++) {
         if (devname[i] >= '0' && devname[i] <= '9') {
             dd->devunit = (int32_t)strtol(devname + i, NULL, 10);
-            *devunit = dd->devunit;
+            *devunit = (uint32_t)dd->devunit;
             return 0;
         }
     }
